+++ CHANGELOG.md
@@ -5,6 +5,11 @@
 
 
 
+### [01/04/2025](https://github.com/vungnv10103/ServerAvatar258)
+
+- [avatar/00:25](https://github.com/vungnv10103/ServerAvatar258/issues) add `module train captcha`, add `template captcha`.
+
+
 ### [30/03/2025](https://github.com/vungnv10103/ServerAvatar258)
 
 - [avatar/22:25](https://github.com/vungnv10103/ServerAvatar258/issues) add `notification`, `farm data`.